Initial commit: add procinfo.py and related files
