+++ .github/copilot-instructions.md
@@ -1,0 +1,3 @@
+<!-- Use this file to provide workspace-specific custom instructions to Copilot. For more details, visit https://code.visualstudio.com/docs/copilot/copilot-customization#_use-a-githubcopilotinstructionsmd-file -->
+
+This project is a Python CLI tool (procinfo) for displaying process information (memory, CPU, disk usage) on Windows using Typer and psutil.
