+++ README.md
@@ -1,13 +1,31 @@
-# python_sp1_procinfo
+# ProcInfo
 
+A Python CLI tool to display process information (memory, CPU, disk usage) for Windows using Typer and psutil.
 
+## Features
+- List top N running processes
+- Show memory, CPU, and disk usage for a selected process
+- User-friendly CLI with prompts
 
-#  Setup 
+## Requirements
+- Python 3.7+
+- Windows OS
+- [psutil](https://pypi.org/project/psutil/)
+- [typer](https://pypi.org/project/typer/)
 
-# create virtual 
-# use Pipfile to configure requirements 
-# update the read me nessary setup or pre-req
+## Installation
+```
+pip install -r requirements.txt
+```
 
+## Usage
+```
+python procinfo.py --platform windows --top-count 5
+```
 
-# Dont directly push to main branch 
-# create feature branch and raise a PR :  please get the knowldge how we can raise a pr from feature branch to master branch 
+## Project Structure
+- `procinfo.py`: Main CLI application
+- `requirements.txt`: Python dependencies
+
+## License
+MIT
